+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v5.0.0...master)
 
  * fixed an issue with using callables in set_context with unsampled transactions (#541, #542)
+ * limited the length of error.culprit to 1024 characters (#491, #543)
 
 ## v5.0.0 
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.2.2...v5.0.0)